rename: UserInfoContainer로 컴포넌트명 변경 및 container, components폴더 생성 후 이동(Next-Study-habjung#56)
